Fix the RedisModule_Free is not 8-byte aligned error on MacOS 14
- Remove mock RedisModule_Free from value_ffi array test
- Reorganize the test files into the tests folder
- Remove unused dependencies in Cargo.toml
