+++ src/redisearch_rs/c_entrypoint/value_ffi/cbindgen.toml
@@ -6,7 +6,7 @@ includes = ["redismodule.h", "query_error.h"]
 
 [parse]
 parse_deps = true
-include = ["c_ffi_utils", "value"]
+include = ["value"]
 
 [export.rename]
 "RsValue" = "RSValue"
