+++ src/redisearch_rs/c_entrypoint/value_ffi/src/array.rs
@@ -107,55 +107,3 @@ pub unsafe extern "C" fn RSValue_ArrayItem(value: *const RsValue, index: u32) ->
         panic!("Expected 'Array' type, got '{}'", value.variant_name())
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::constructors::RSValue_NewNumber;
-    use crate::getters::RSValue_Number_Get;
-    use crate::shared::RSValue_DecrRef;
-
-    #[unsafe(no_mangle)]
-    pub static mut RSDummyContext: *mut ffi::RedisModuleCtx = std::ptr::null_mut();
-
-    #[unsafe(no_mangle)]
-    pub unsafe extern "C" fn RedisModule_Free() {
-        unimplemented!("RedisModule_Free")
-    }
-
-    #[unsafe(no_mangle)]
-    pub unsafe extern "C" fn RedisModule_FreeString() {
-        unimplemented!("RedisModule_FreeString")
-    }
-
-    #[test]
-    fn test_array() {
-        unsafe {
-            let array = RSValue_NewArrayBuilder(3);
-            let one = RSValue_NewNumber(1.0);
-            let two = RSValue_NewNumber(2.0);
-            let three = RSValue_NewNumber(3.0);
-            array.add(0).write(one);
-            array.add(1).write(two);
-            array.add(2).write(three);
-
-            let array = RSValue_NewArrayFromBuilder(array, 3);
-
-            assert_eq!(RSValue_ArrayLen(array), 3);
-
-            let val = RSValue_ArrayItem(array, 0);
-            let num = RSValue_Number_Get(val);
-            assert_eq!(1.0, num);
-
-            let val = RSValue_ArrayItem(array, 1);
-            let num = RSValue_Number_Get(val);
-            assert_eq!(2.0, num);
-
-            let val = RSValue_ArrayItem(array, 2);
-            let num = RSValue_Number_Get(val);
-            assert_eq!(3.0, num);
-
-            RSValue_DecrRef(array);
-        }
-    }
-}
